refactor: update files naming to kebab-case
